CATL-1804: Run activity setup on install and update
